refactor: Ajustar ruta de módulo plantilla a nuevo esquema src/application de migración udistrital/sisifo_documentacion#522
