+++ src/application/plantilla/plantilla.module.ts
@@ -3,7 +3,7 @@ import { PlantillaController } from './plantilla.controller';
 import { PlantillaService } from './services/plantilla.service';
 import { HttpModule } from '@nestjs/axios';
 import { PlantillaPlanTrabajoService } from './services/plantilla-plan-trabajo.service';
-import { PlantillaUtilsService } from '../utils/plantilla.utils';
+import { PlantillaUtilsService } from '../../utils/plantilla.utils';
 import { PlantillaSolicitudInformacionService } from './services/plantilla-solicitud-informacion.service';
 import { PlantillaCartaPresentacionService } from './services/plantilla-carta-presentacion.service';
 import { PlantillaProgramaAuditoriaService } from './services/plantilla-programa-auditoria.service';
