+++ src/application/plantilla/services/plantilla-carta-presentacion.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@nestjs/common';
 import * as moment from 'moment';
 import 'moment/locale/es';
-import { PlantillaUtilsService } from '../../utils/plantilla.utils';
+import { PlantillaUtilsService } from '../../../utils/plantilla.utils';
 import { environment } from 'src/config/configuration';
 
 @Injectable()
